requirements.txt
FastAPI
uvicorn
httpx
pydantic
python-dotenv
requests
jinja2